bindings/python/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "fluss-python"
description = "Python bindings for Fluss"
authors = [{name = "Fluss Team"}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
]
dependencies = [
    "pyarrow>=10.0.0",
    "pandas>=1.5.0",
    "duckdb>=0.9.0",
    "pylance>=0.33.0",
    "tqdm>=4.67.1",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/naivedogger/fluss-rs"
Repository = "https://github.com/naivedogger/fluss-rs"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0",
    "black>=22.0",
    "ruff>=0.1.0",
]

[tool.maturin]
python-source = "python"
module-name = "fluss_python"
features = ["pyo3/extension-module"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
